+++ gateway.tf
@@ -6,10 +6,10 @@ resource "aws_api_gateway_rest_api" "main" {
   minimum_compression_size = var.api_gateway_minimum_compression_size
   tags                     = module.labels.tags
 
-  binary_media_types = [
+  binary_media_types = concat([
     "application/zip",
-    "application/octet-stream"
-  ]
+    "application/octet-stream",
+  ], var.api_gateway_customizations_binary_types)
 
   endpoint_configuration {
     types = ["EDGE"]